+++ src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import Row from "./Row";
+import Row from "./components/Row";
 import Section from "./Section";
 
 function App() {
